backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0
httpx>=0.26.0

# Firebase Admin
firebase-admin>=6.4.0

# Whisper STT (local)
openai-whisper>=20231117
torch>=2.1.0

# AI LLMs
groq>=0.4.0
google-generativeai>=0.3.0

# Agora WebRTC
agora-token-builder>=1.0.0

# Utils
python-dotenv>=1.0.0
aiofiles>=23.2.0
